Updated MPMICE

+++ Vaango/src/CCA/Components/MPMICE/MPMICE.cc
@@ -1217,7 +1217,7 @@ MPMICE::coarsenVariableCC(const ProcessorGroup*,
 
       Level::selectType finePatches;
       coarsePatch->getFineLevelPatches(finePatches);
-      for (int i = 0; i < finePatches.size(); i++) {
+      for (size_t i = 0; i < finePatches.size(); i++) {
         const Patch* finePatch = finePatches[i];
 
         IntVector cl, ch, fl, fh;
@@ -1413,7 +1413,7 @@ MPMICE::coarsenVariableNC(const ProcessorGroup*,
 
       Level::selectType finePatches;
       coarsePatch->getFineLevelPatches(finePatches);
-      for (int i = 0; i < finePatches.size(); i++) {
+      for (size_t i = 0; i < finePatches.size(); i++) {
         const Patch* finePatch = finePatches[i];
 
         IntVector cl, ch, fl, fh;
